add delete method in usuario, alumno, tecnico and personal classes and add documentation

+++ proyecto_satapp_Carlos_Rafa/src/main/java/com/example/proyecto_satapp_Carlos_Rafa/services/AlumnoService.java
@@ -72,8 +72,13 @@ public class AlumnoService {
     }
 
 
-    public void delete(Long id) {
+    public void deleteById(Long id) {
+        Optional<Alumno> alumnoOp = alumnoRepository.findById(id);
+
+        if (alumnoOp.isEmpty()) {
+            throw new EntityNotFoundException("Alumno no encontrado");
+        }
+
         alumnoRepository.deleteById(id);
     }
-
 }
